feat, refactor: Updating code and handling error

+++ coders/codexion.c
@@ -6,7 +6,7 @@
 /*   By: nyramana <nyramana@student.42antananarivo  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/07/07 11:48:24 by nyramana          #+#    #+#             */
-/*   Updated: 2026/07/07 22:05:45 by nyramana         ###   ########.fr       */
+/*   Updated: 2026/07/11 08:12:48 by nyramana         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,8 @@ int	main(int argc, char **argv)
 		return (1);
 	if (!init_dongles(&all) || !init_all(&all))
 		return (1);
-	init_coders(&all);
+	if (!init_coders(&all))
+		return (1);
 	destroy_all(&all);
 	return (0);
 }
